posix: validate mode argument in posix_access

+++ ext/posix/tests/posix_access_flags.phpt
@@ -45,10 +45,15 @@ if (posix_access($testfile, POSIX_R_OK | POSIX_W_OK)) {
 if (posix_access($testfile, POSIX_F_OK)) {
     echo "File exists OK\n";
 }
+
 ?>
 --CLEAN--
 <?php
-unlink($testfile);
+$dir = __DIR__;
+$testfile = "$dir/testfile.txt";
+if (file_exists($testfile)) {
+    unlink($testfile);
+}
 ?>
 --EXPECTF--
 posix_access(): Argument #2 ($flags) must be a bitmask of POSIX_F_OK, POSIX_R_OK, POSIX_W_OK, and POSIX_X_OK
